[CRUD_Producto] CRUD finalizado, con 2/5 tests funcionales
